@@ -4,8 +4,9 @@ import android.app.Activity
 import android.content.Intent
 import android.os.Environment
 import android.provider.MediaStore
-import android.support.v4.content.FileProvider
+import androidx.core.content.FileProvider
 import java.io.File
+import java.io.FileInputStream
 
 /**
  * @author applexis
@@ -27,13 +28,13 @@ class MediaPicker(private val activity: Activity) {
         private var tmpDestinationPath = ""
     }
 
-    private val OUTPUT_FILENAME_DIR = File(Environment.getExternalStorageDirectory(), Environment.DIRECTORY_MOVIES)
     private lateinit var callback: (source: Int, path: String) -> Unit
     var dialog: BaseMPDialog? = null
         set(value) {
             field = value
             dialog?.init(this)
         }
+    private val outputFile: File = File(activity.cacheDir, "tmp.jpg")
 
     fun show(callback: (source: Int, path: String) -> Unit) {
         this.callback = callback
@@ -45,19 +46,9 @@ class MediaPicker(private val activity: Activity) {
     }
 
     internal fun pickPhotoFromCamera() {
-        val name = currentDatetimeFilename()
-        if (!OUTPUT_FILENAME_DIR.exists()) {
-            OUTPUT_FILENAME_DIR.mkdirs()
-        }
-        val destination = File(
-                Environment.getExternalStorageDirectory(),
-                Environment.DIRECTORY_MOVIES + "/" + name + ".jpg"
-        )
-
-        tmpDestinationPath = destination.absolutePath
-
+        tmpDestinationPath = outputFile.absolutePath
         val uriPhoto = FileProvider.getUriForFile(activity,
-                activity.packageName + ".provider", destination)
+                activity.packageName + ".provider", outputFile)
         val intent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
         intent.putExtra(MediaStore.EXTRA_OUTPUT, uriPhoto)
         dismiss()
@@ -65,19 +56,9 @@ class MediaPicker(private val activity: Activity) {
     }
 
     internal fun pickVideoFromCamera() {
-        val nameVideo = currentDatetimeFilename()
-        if (!OUTPUT_FILENAME_DIR.exists()) {
-            OUTPUT_FILENAME_DIR.mkdirs()
-        }
-        val destination = File(
-                Environment.getExternalStorageDirectory(),
-                Environment.DIRECTORY_MOVIES + "/" + nameVideo + ".mp4"
-        )
-
-        tmpDestinationPath = destination.absolutePath
-
+        tmpDestinationPath = outputFile.absolutePath
         val uriVideo = FileProvider.getUriForFile(activity,
-                activity.packageName + ".provider", destination)
+                activity.packageName + ".provider", outputFile)
         val intentVideo = Intent(MediaStore.ACTION_VIDEO_CAPTURE)
         intentVideo.putExtra(MediaStore.EXTRA_OUTPUT, uriVideo)
         dismiss()
@@ -85,14 +66,14 @@ class MediaPicker(private val activity: Activity) {
     }
 
     internal fun pickPhotoFromGallery() {
-        val selectIntent = Intent(Intent.ACTION_PICK, android.provider.MediaStore.Images.Media.EXTERNAL_CONTENT_URI)
+        val selectIntent = Intent(Intent.ACTION_PICK, MediaStore.Images.Media.EXTERNAL_CONTENT_URI)
         selectIntent.type = "image/*"
         dismiss()
         activity.startActivityForResult(selectIntent, LIBRARY_PHOTO_REQUEST)
     }
 
     internal fun pickVideoFromGallery() {
-        val selectIntentVid = Intent(Intent.ACTION_PICK, android.provider.MediaStore.Video.Media.EXTERNAL_CONTENT_URI)
+        val selectIntentVid = Intent(Intent.ACTION_PICK, MediaStore.Video.Media.EXTERNAL_CONTENT_URI)
         selectIntentVid.type = "video/*"
         dismiss()
         activity.startActivityForResult(selectIntentVid, LIBRARY_VIDEO_REQUEST)
@@ -103,17 +84,20 @@ class MediaPicker(private val activity: Activity) {
             when (requestCode) {
                 LIBRARY_PHOTO_REQUEST, LIBRARY_VIDEO_REQUEST -> {
                     val selectedImageUri = data?.data
-                    try {
-                        val filePathColumn = arrayOf(MediaStore.Video.Media.DATA)
-                        val cursor = activity.contentResolver.query(selectedImageUri, filePathColumn, null, null, null)
-                        cursor!!.moveToFirst()
-
-                        val columnIndex = cursor.getColumnIndex(filePathColumn[0])
-                        val resultPath = cursor.getString(columnIndex)
-                        cursor.close()
-                        callback(requestCode - CODE_ADD, resultPath)
-                    } catch (e: Exception) {
-                        e.printStackTrace()
+                    if (selectedImageUri != null) {
+                        try {
+                            val fileDescriptor = activity.contentResolver.openFileDescriptor(selectedImageUri, "r")?.fileDescriptor
+                            if (fileDescriptor != null) {
+                                FileInputStream(fileDescriptor).use { input ->
+                                    outputFile.outputStream().use { output ->
+                                        input.copyTo(output)
+                                    }
+                                }
+                            }
+                            callback(requestCode - CODE_ADD, outputFile.absolutePath)
+                        } catch (e: Exception) {
+                            e.printStackTrace()
+                        }
                     }
                 }
 
